Move: product option, image 관련 로직을 vendor 도메인으로 이동
